# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My restaurant Project CRUD",
        "description": "My User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
        {
            "name": "Restaurants",
            "description": "API for users in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get all restaurants",
                "description": "Get all restaurants",
                "operationId": "getRestaurants",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Restaurant"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        }, 
        "/restaurant": {
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Add a new restaurant",
                "description": "Add a new restaurant",
                "operationId": "addRestaurant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Restaurant",
                        "description": "Restaurant that we want to create",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Restaurant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Restaurant"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        }, 
        "/restaurant/{id}": {
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Delete a restaurant",
                "description": "Delete a restaurant",
                "operationId": "deleteRestaurant",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of restaurant to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Restaurant"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Restaurant not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Update a restaurant",
                "description": "Update a restaurant",
                "operationId": "updateRestaurant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of restaurant to update",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "Restaurant",
                        "description": "Restaurant that we want to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Restaurant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Restaurant"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Restaurant not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Restaurant": {
            "required": ["id", "name"],
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}